broke submit message button to its own component, WIP filter input field component not functioning
