+++ lib/components/Application.js
@@ -6,11 +6,13 @@ import SubmitMessage from './submit-message.js';
 import ClearInput from './clear-input.js';
 import LogInButton from './login-button.js';
 import LogOutButton from './logout-button.js';
+import FilterInputField from './filter-input-field.js';
+import ChatMaster from './chatMaster.js';
 import SortUp from './sort-up.js';
 import SortDown from './sort-down.js';
 import moment from 'moment';
 
-//TODO: for refactor tractor: break out components as noted above; try to refactor JS if statements; get closer match to specs if time permits 
+//TODO: for refactor tractor: break out components as noted above; try to refactor JS if statements; get closer match to specs if time permits
 
 export default class Application extends Component {
   constructor() {
@@ -314,6 +316,7 @@ export default class Application extends Component {
 
   updateSearch(event) {
     let searchValue = {search: event.target.value.substr(0, 20)};
+    console.log(event);
     this.setState(searchValue);
     this.findMatchingMessages(searchValue);
   }
@@ -324,11 +327,9 @@ export default class Application extends Component {
       <div onLoad={this.initFirebase()}>
         <header>
           <h2 id="page-main-title">Shoot the Breeze</h2>
-          <input id="filter-input-field"
-                 type="text"
-                 placeholder="Filter"
-                 value={this.state.search}
-                 onChange={this.updateSearch.bind(this)} />
+          <FilterInputField
+          handleChange = {() => this.updateSearch.bind(this)}
+          value={this.state.search}/>
           <div className="buttons-container-header">
             <SortUp handleClick={()=>this.sortUp()}/>
             <SortDown handleClick={()=>this.sortDown()}/>
@@ -369,62 +370,5 @@ export default class Application extends Component {
 
 } //end of Application
 
-class ChatMaster extends React.Component { //TODO: put this in own file
-  constructor(props) {
-    super(props);
-    this.state = {
-      filteredMssgs: []
-    }
-
-  } //end of constructor
-
-  filterMessages(user){
-    let newMssgs = this.props.messages.filter((message)=>{
-      return message.user === user;
-    });
-    this.wipeOutAllMessagesFromScreen();
-    this.setState({filteredMssgs: newMssgs});
-  }
-
-  showMessages(messages){
-    return(
-      <div className="each-message-card">
-        <p><span className="each-message-date">{messages.time}</span><span className="each-message-user-name">{messages.user}</span> <a href={"mailto:" + messages.email}>({messages.email})</a></p>
-        <p className="each-message-body-text">{messages.body}</p><br/>
-      </div>
-    )
-  }
-
-  wipeOutAllMessagesFromScreen(){
-    document.querySelector('.messages').textContent = '';
-  }
-
-  render () {
-    const { messages, filterUser } = this.props;
-    const users = messages.map(message => message.user);
-    const uniqueUsers = _.uniq(users);
-
-    return (
-      <div>
-        <div className="message-output-container">
-          <ul>
-            <li className="messages">{messages.map(this.showMessages)}</li>
-            <li>{this.state.filteredMssgs.map(this.showMessages)}</li>
-          </ul>
-        </div>
-        <div className="user-info-container">
-          <div className="users-and-filter-buttons-container">
-            <h2>Users Present in the Chat Room:</h2>
-            <p>
-              {uniqueUsers.map(user =>
-                <button id={user.substr(0, 3)} className="users-present-button" onClick={() => this.filterMessages(user)}>{user}</button>
-              )}
-            </p>
-          </div>
-        </div>
-      </div> //end of main div
-    );
-  }
-} //end of ChatMaster
 
 module.exports = Application;
